rename: separate TodoContent component from TodoList
